Websocket: Move connection type into its file
